created StatTable class to abstract out the HTML in the main app
ended up doing lots of refactoring to keep things organized and reduce the file sizes
